Use .indexOf to find the end of a ruby's body
I was previously using a helper
function given by markdown-it,
which can find matching square
brackets, but a much simpler
.indexOf also works.

That helper function was only useful before because I was
also recursively formatting the
body, and that function handled
escaping properly. However, I
got rid of the recursive formatting
a few commits ago, so now it
doesn't matter.